+++ packages/mail/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @lunora/mail [1.0.0-alpha.80](https://github.com/anolilab/lunora/compare/@lunora/mail@1.0.0-alpha.79...@lunora/mail@1.0.0-alpha.80) (2026-09-12)
+
+
+### Dependencies
+
+* **@lunora/errors:** upgraded to 1.0.0-alpha.36
+
+## @lunora/mail [1.0.0-alpha.79](https://github.com/anolilab/lunora/compare/@lunora/mail@1.0.0-alpha.78...@lunora/mail@1.0.0-alpha.79) (2026-09-12)
+
 ## @lunora/mail [1.0.0-alpha.78](https://github.com/anolilab/lunora/compare/@lunora/mail@1.0.0-alpha.77...@lunora/mail@1.0.0-alpha.78) (2026-09-11)
 
 ## @lunora/mail [1.0.0-alpha.77](https://github.com/anolilab/lunora/compare/@lunora/mail@1.0.0-alpha.76...@lunora/mail@1.0.0-alpha.77) (2026-09-11)