+++ packages/notify/CHANGELOG.md
@@ -1,3 +1,18 @@
+## @lunora/notify [1.0.0-alpha.56](https://github.com/anolilab/lunora/compare/@lunora/notify@1.0.0-alpha.55...@lunora/notify@1.0.0-alpha.56) (2026-09-12)
+
+### Bug Fixes
+
+* **notify:** do not retry a partially delivered push ([#715](https://github.com/anolilab/lunora/issues/715)) ([7a28260](https://github.com/anolilab/lunora/commit/7a282608f86f27f0bbcd424d5fef9418317d57d0))
+
+## @lunora/notify [1.0.0-alpha.55](https://github.com/anolilab/lunora/compare/@lunora/notify@1.0.0-alpha.54...@lunora/notify@1.0.0-alpha.55) (2026-09-12)
+
+
+### Dependencies
+
+* **@lunora/errors:** upgraded to 1.0.0-alpha.36
+
+## @lunora/notify [1.0.0-alpha.54](https://github.com/anolilab/lunora/compare/@lunora/notify@1.0.0-alpha.53...@lunora/notify@1.0.0-alpha.54) (2026-09-12)
+
 ## @lunora/notify [1.0.0-alpha.53](https://github.com/anolilab/lunora/compare/@lunora/notify@1.0.0-alpha.52...@lunora/notify@1.0.0-alpha.53) (2026-09-11)
 
 ## @lunora/notify [1.0.0-alpha.52](https://github.com/anolilab/lunora/compare/@lunora/notify@1.0.0-alpha.51...@lunora/notify@1.0.0-alpha.52) (2026-09-11)