+++ packages/x402/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @lunora/x402 [1.0.0-alpha.69](https://github.com/anolilab/lunora/compare/@lunora/x402@1.0.0-alpha.68...@lunora/x402@1.0.0-alpha.69) (2026-09-12)
+
+
+### Dependencies
+
+* **@lunora/errors:** upgraded to 1.0.0-alpha.36
+
+## @lunora/x402 [1.0.0-alpha.68](https://github.com/anolilab/lunora/compare/@lunora/x402@1.0.0-alpha.67...@lunora/x402@1.0.0-alpha.68) (2026-09-12)
+
 ## @lunora/x402 [1.0.0-alpha.67](https://github.com/anolilab/lunora/compare/@lunora/x402@1.0.0-alpha.66...@lunora/x402@1.0.0-alpha.67) (2026-09-11)
 
 ## @lunora/x402 [1.0.0-alpha.66](https://github.com/anolilab/lunora/compare/@lunora/x402@1.0.0-alpha.65...@lunora/x402@1.0.0-alpha.66) (2026-09-11)